add test and make container work unprivileged

+++ api/nivo_api/cli/database.py
@@ -3,6 +3,7 @@ from nivo_api.core.db.models.nivo import *
 from nivo_api.core.db.models.bra import *
 
 
-def create_schema_and_table(drop: Boolean) -> None:
-    metadata.drop_all(db_engine)
+def create_schema_and_table(drop: bool) -> None:
+    if drop:
+        metadata.drop_all(db_engine)
     metadata.create_all(db_engine)
